+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.12.0
+
+* **BREAKING** The redux utilities are no longer part of the core package, but need to be imported from `mobx/middleware/redux`.
+
 # 0.11.0
 
 ### Breaking changes